+++ Backend/MessageService/include/messageservice/interfaces/IController.h
@@ -10,7 +10,7 @@ class GetMessagePack;
 class IController {
   public:
     virtual std::vector<Message> getMessages(const GetMessagePack&) = 0;
-    virtual std::vector<MessageStatus> getMessagesStatus(const std::vector<Message>&, int receiver_id) = 0;
+    virtual std::vector<MessageStatus> getMessagesStatus(const std::vector<Message>&, long long receiver_id) = 0;
 };
 
 #endif // ICONTROLLER_H
